lint

+++ src/animatediff/adw/schema.py
@@ -14,7 +14,7 @@ class TStatusEnum(str, enum.Enum):
 
 class TSubtask(PtBaseModel):
     description: str = ""
-    completed: int = 0
+    completed: float = 0.0
     total: int = 100
     status: TStatusEnum = TStatusEnum.PENDING
 
@@ -22,7 +22,7 @@ class TSubtask(PtBaseModel):
 class TPipeline(PtBaseModel):
     pid: int
     status: TStatusEnum = TStatusEnum.PENDING
-    completed: int = 0
+    completed: float = 0.0
     total: int = 100
     subtasks: list[TSubtask] = pt.Field(default_factory=list)
     video_path: str = ""
